Create methodWrappers file and move helpers over to it
This includes helpers like plugin.CName, plugin.Name, etc.
