+++ plugin/methodWrappers_test.go
@@ -1,0 +1,85 @@
+package plugin
+
+import (
+	"testing"
+	"time"
+
+	"github.com/puppetlabs/wash/datastore"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/suite"
+)
+
+type MethodWrappersTestSuite struct {
+	suite.Suite
+}
+
+func (suite *MethodWrappersTestSuite) SetupSuite() {
+	SetTestCache(datastore.NewMemCache())
+}
+
+func (suite *MethodWrappersTestSuite) TearDownSuite() {
+	UnsetTestCache()
+}
+
+func (suite *MethodWrappersTestSuite) TestName() {
+	e := newmethodWrappersTestsMockEntry("foo")
+	suite.Equal(Name(e), "foo")
+}
+
+func (suite *MethodWrappersTestSuite) TestCName() {
+	e := newmethodWrappersTestsMockEntry("foo/bar/baz")
+	suite.Equal("foo#bar#baz", CName(e))
+
+	e.SetSlashReplacer(':')
+	suite.Equal("foo:bar:baz", CName(e))
+}
+
+func (suite *MethodWrappersTestSuite) TestID() {
+	e := newmethodWrappersTestsMockEntry("foo/bar")
+
+	e.SetTestID("")
+	suite.Panics(
+		func() { ID(e) },
+		"plugin.ID: entry foo (cname foo#bar) has no ID",
+	)
+
+	e.SetTestID("/foo/bar")
+	suite.Equal(ID(e), "/foo/bar")
+}
+
+type methodWrappersTestsMockEntry struct {
+	EntryBase
+	mock.Mock
+}
+
+func (e *methodWrappersTestsMockEntry) Schema() *EntrySchema {
+	return nil
+}
+
+func newmethodWrappersTestsMockEntry(name string) *methodWrappersTestsMockEntry {
+	e := &methodWrappersTestsMockEntry{
+		EntryBase: NewEntry(name),
+	}
+	e.SetTestID("id")
+	e.DisableDefaultCaching()
+
+	return e
+}
+
+func (suite *MethodWrappersTestSuite) TestAttributes() {
+	e := newmethodWrappersTestsMockEntry("mockEntry")
+	e.attr = EntryAttributes{}
+	e.attr.SetCtime(time.Now())
+	suite.Equal(e.attr, Attributes(e))
+}
+
+func (suite *MethodWrappersTestSuite) TestPrefetched() {
+	e := newmethodWrappersTestsMockEntry("mockEntry")
+	suite.False(IsPrefetched(e))
+	e.Prefetched()
+	suite.True(IsPrefetched(e))
+}
+
+func TestMethodWrappers(t *testing.T) {
+	suite.Run(t, new(MethodWrappersTestSuite))
+}
